backend/seguridad/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose
pydantic[email]
pydantic-settings
